Continued fixing wannings

+++ init.c
@@ -20,7 +20,7 @@ void* Mem_Init(int sizeOfRegion){
 	}
 	//Pass 0 to default to CHUNK
 	size_t size = sizeOfRegion == 0 ?
-		(size_t)CHUNK : sizeOfRegion;
+		(size_t)CHUNK : (size_t)sizeOfRegion;
 	return extend_extendInit(size);
 	// |h|H|h|h|h| | | | | | |
 }
